test(plugins): split Gateway mutation coverage
